% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{albertinedisparue}
\alias{albertinedisparue}
\title{Marcel Proust's novel "Albertine disparue"}
\format{A character vector with 392 elements}
\source{
\url{https://beq.ebooksgratuits.com/auteurs/Proust/proust.htm}
}
\usage{
albertinedisparue
}
\description{
A dataset containing the first part of Marcel Proust's "Du côté de chez Swann". 
This text has been downloaded and converted to txt from "La Bibliothèque électronique du Québec".
}
\keyword{datasets}
