% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_novels.R
\docType{data}
\name{alombredesjeunesfillesenfleurs}
\alias{alombredesjeunesfillesenfleurs}
\title{Marcel Proust's novel "À l’ombre des jeunes filles en fleurs"}
\format{A tibble with text, book, volume, and year}
\source{
<https://fr.wikisource.org/wiki/%C3%80_l%E2%80%99ombre_des_jeunes_filles_en_fleurs>
}
\usage{
alombredesjeunesfillesenfleurs
}
\description{
A dataset containing Marcel Proust's "À l’ombre des jeunes filles en fleurs". 
This text has been downloaded from WikiSource.
}
\keyword{datasets}
