% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{alombredesjeunesfillesenfleurs1}
\alias{alombredesjeunesfillesenfleurs1}
\title{The first part of Marcel Proust's novel "À l’ombre des jeunes filles en fleurs"}
\format{A character vector with 373 elements}
\source{
\url{https://beq.ebooksgratuits.com/auteurs/Proust/proust.htm}
}
\usage{
alombredesjeunesfillesenfleurs1
}
\description{
A dataset containing the first part of Marcel Proust's "À l’ombre des jeunes filles en fleurs". 
This text has been downloaded and converted to txt from "La Bibliothèque électronique du Québec".
}
\keyword{datasets}
