% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{alombredesjeunesfillesenfleurs3}
\alias{alombredesjeunesfillesenfleurs3}
\title{The third part of Marcel Proust's novel "À l’ombre des jeunes filles en fleurs"}
\format{A character vector with 329 elements}
\source{
\url{https://beq.ebooksgratuits.com/auteurs/Proust/proust.htm}
}
\usage{
alombredesjeunesfillesenfleurs3
}
\description{
A dataset containing the first part of Marcel Proust's "Du côté de chez Swann". 
This text has been downloaded and converted to txt from "La Bibliothèque électronique du Québec".
}
\keyword{datasets}
