% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_novels.R
\docType{data}
\name{ducotedechezswann}
\alias{ducotedechezswann}
\title{Marcel Proust's novel "Du côté de chez Swann"}
\format{A tibble with text, book, volume, and year}
\source{
<https://fr.wikisource.org/wiki/Du_c%C3%B4t%C3%A9_de_chez_Swann>
}
\usage{
ducotedechezswann
}
\description{
A dataset containing Marcel Proust's "Du côté de chez Swann". 
This text has been downloaded from WikiSource.
}
\keyword{datasets}
