% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{ducotedechezswann1}
\alias{ducotedechezswann1}
\title{The first part of Marcel Proust's novel "Du côté de chez Swann"}
\format{A character vector with 645 elements}
\source{
\url{https://beq.ebooksgratuits.com/auteurs/Proust/proust.htm}
}
\usage{
ducotedechezswann1
}
\description{
A dataset containing the first part of Marcel Proust's "Du côté de chez Swann". 
This text has been downloaded and converted to txt from "La Bibliothèque électronique du Québec".
}
\keyword{datasets}
