% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{ducotedechezswann2}
\alias{ducotedechezswann2}
\title{The second part of Marcel Proust's novel "Du côté de chez Swann"}
\format{A character vector 614 elements}
\source{
\url{https://beq.ebooksgratuits.com/auteurs/Proust/proust.htm}
}
\usage{
ducotedechezswann2
}
\description{
A dataset containing the second part of Marcel Proust's "Du côté de chez Swann". 
This text has been downloaded and converted to txt from "La Bibliothèque électronique du Québec".
}
\keyword{datasets}
