% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{laprisonniere1}
\alias{laprisonniere1}
\title{The first part of Marcel Proust's novel "La Prisonnière"}
\format{A character vector with 389 elements}
\source{
\url{https://beq.ebooksgratuits.com/auteurs/Proust/proust.htm}
}
\usage{
laprisonniere1
}
\description{
A dataset containing the first part of Marcel Proust's "À l’ombre des jeunes filles en fleurs". 
This text has been downloaded and converted to txt from "La Bibliothèque électronique du Québec".
}
\keyword{datasets}
