% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_novels.R
\docType{data}
\name{lecotedeguermantes}
\alias{lecotedeguermantes}
\title{Marcel Proust's novel "Le côté de Guermantes"}
\format{A tibble with text, book, volume, and year}
\usage{
lecotedeguermantes
}
\description{
A dataset containing Marcel Proust's "À l’ombre des jeunes filles en fleurs". 
This text has been downloaded from WikiSource.
}
\keyword{datasets}
