% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_novels.R
\docType{data}
\name{letempretrouve}
\alias{letempretrouve}
\title{Marcel Proust's novel "Le temps retrouvé"}
\format{A tibble with text, book, volume, and year.}
\usage{
letempretrouve
}
\description{
A dataset containing Marcel Proust's "Le temps retrouvé". 
This text has been downloaded from WikiSource.
}
\keyword{datasets}
