% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{letempretrouve2}
\alias{letempretrouve2}
\title{The second part of Marcel Proust's novel "Le temps retrouvé"}
\format{A character vector with 270 elements}
\source{
\url{https://beq.ebooksgratuits.com/auteurs/Proust/proust.htm}
}
\usage{
letempretrouve2
}
\description{
A dataset containing the first part of Marcel Proust's "Du côté de chez Swann". 
This text has been downloaded and converted to txt from "La Bibliothèque électronique du Québec".
}
\keyword{datasets}
