% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_detect_calendar.R
\name{pr_detect_days}
\alias{pr_detect_days}
\title{Detect french days}
\usage{
pr_detect_days(df, col)
}
\arguments{
\item{df}{a dataframe}

\item{col}{the column containing the text}
}
\value{
a tibble with the number of days detected by the algo
}
\description{
Detect the name of the days (in French)
}
\examples{
a <- data.frame(jours = c("C'est lundi 1er mars et mardi 2", 
"Et mercredi 3", "Il est revenu jeudi."))
pr_detect_days(a, jours)
}
