% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_detect_pronouns.R
\name{pr_detect_pro}
\alias{pr_detect_pro}
\title{Detect pronoums}
\usage{
pr_detect_pro(df, col, verbose = FALSE)
}
\arguments{
\item{df}{a dataframe}

\item{col}{the column containing the text}

\item{verbose}{wether or not to return the list of pronouns. Defaults is FALSE.}
}
\value{
a dataframe
}
\description{
Detect the pronouns from a text.
}
\examples{
library(proustr)
a <- proust_books()[1,] 
pr_detect_pro(a, text, verbose = TRUE)
pr_detect_pro(a, text)
}
