% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_norm.R
\name{pr_keep_only_alnum}
\alias{pr_keep_only_alnum}
\title{Remove non alnum elements}
\usage{
pr_keep_only_alnum(text, replacement = " ")
}
\arguments{
\item{text}{a vector}

\item{replacement}{what to replace the non alnum with. Defaut is " ".}
}
\value{
a vector
}
\description{
Remove non alnum elements
}
\examples{
pr_keep_only_alnum("neuilly-en-thelle")
}
