% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_stem.R
\name{pr_stem_sentences}
\alias{pr_stem_sentences}
\title{Stem a dataframe containing a column with sentences}
\usage{
pr_stem_sentences(df, col, language = "french")
}
\arguments{
\item{df}{the data.frame containing the text}

\item{col}{the column with the text}

\item{language}{the language of the text. Defaut is french. See SnowballC::getStemLanguages() function for a list of supported languages.}
}
\value{
a tibble
}
\description{
Implementation of the {SnowballC} stemmer. Note that punctuation and capital letters 
are removed when processing.
}
\examples{
a <- proustr::laprisonniere[1:10,]
pr_stem_sentences(a, text)

}
