% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_stem.R
\name{pr_stem_words}
\alias{pr_stem_words}
\title{Stem a dataframe containing a column with words}
\usage{
pr_stem_words(df, col, language = "french")
}
\arguments{
\item{df}{the data.frame containing the sentences}

\item{col}{the column with the sentences}

\item{language}{the language of the words Defaut is french. See SnowballC::getStemLanguages() function for a list of supported languages.}
}
\value{
a tibble
}
\description{
Implementation of the {SnowballC} stemmer. Note that punctuation and capitals letters 
are also removed.
}
\examples{
a <- data.frame(words = c("matin", "heure", "fatigué","sonné","lois", "tests","fusionner"))
pr_stem_words(a, words)

}
