% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_norm.R
\name{pr_unacent}
\alias{pr_unacent}
\title{Remove accents}
\usage{
pr_unacent(text)
}
\arguments{
\item{text}{a vector}
}
\value{
a vector
}
\description{
Remove accents from a character vector
}
\examples{
pr_unacent("du chêne")
}
