% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_books.R
\name{proust_books}
\alias{proust_books}
\title{Tidy data frame of Marcel Proust's 7 novels from La Recherche}
\usage{
proust_books()
}
\value{
A tibble with three columns: \code{text}, \code{book}, and \code{volume}.
}
\description{
Returns a tidy data frame of Marcel Proust's 7 novels from À la recherche du temps
perdu. The data frame contains one column with the text, one with the book, and one with the 
volume.
}
\examples{

#Creates the tibble 
proust <- proust_books()
 

}
