% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_datasets.R
\docType{data}
\name{proust_char}
\alias{proust_char}
\title{Characters from "À la recherche du temps perdu"}
\format{A tibble with their name}
\source{
\url{http://proust-personnages.fr/?page_id=10254}
}
\usage{
proust_char
}
\description{
A dataset containing Marcel Proust's characters from "À la recherche du temps perdu"  and 
their frequency in each book. 
This dataset has been downloaded from proust-personnages.
}
\keyword{datasets}
