% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_random.R
\name{proust_random}
\alias{proust_random}
\title{Create a Random Proust extract}
\usage{
proust_random(count = 1, collapse = TRUE)
}
\arguments{
\item{count}{the number of line you want to randomly extract and paste.}

\item{collapse}{if FALSE, the output will be a tibble. Default is TRUE, a character vector.}
}
\value{
a character vector
}
\description{
Create your own flavor of Proust with this random extractor.
}
\examples{
proust_random(3)
}
