% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_sentiments.R
\name{proust_sentiments}
\alias{proust_sentiments}
\title{Old sentiment lexicon}
\usage{
proust_sentiments(type = c("polarity", "score"))
}
\value{
a tibble
}
\description{
This function will be removed in next proustr version.
}
