% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_stop_words.R
\name{proust_stopwords}
\alias{proust_stopwords}
\title{Stop Words}
\source{
\url{https://raw.githubusercontent.com/stopwords-iso/stopwords-fr/master/stopwords-fr.txt}

\url{http://www.ranks.nl/stopwords/french}

\url{http://www.naunaute.com/liste-stop-words-francais-393}

\url{https://sites.google.com/site/kevinbouge/stopwords-lists/stopwords_fr.txt?attredirects=0&d=1}
}
\usage{
proust_stopwords()
}
\value{
a tibble with stopwords
}
\description{
Stop words concatenated from various web sources.
}
\examples{
proust_stopwords()
}
