% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{package}
\name{proustr}
\alias{proustr}
\alias{proustr-package}
\alias{proustr-package}
\title{Proust's À la recherche du temps perdu}
\description{
Texts from Marcel Proust's "A La Recherche Du Temps Perdu".
}
