% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_datasets.R
\docType{data}
\name{sentiments_polarity}
\alias{sentiments_polarity}
\title{Sentiment lexicon with polarity}
\format{A tibble}
\source{
\url{http://www.lirmm.fr/~abdaoui/FEEL.html}
}
\usage{
sentiments_polarity
}
\description{
Dataset from: Amine Abdaoui, Jérôme Azé, Sandra Bringay et Pascal Poncelet. FEEL: French Expanded Emotion Lexicon. Language Resources and Evaluation, LRE 2016, pp 1-23.
}
\keyword{datasets}
