% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_datasets.R
\docType{data}
\name{sentiments_score}
\alias{sentiments_score}
\title{Sentiment lexicon with score}
\format{A tibble}
\source{
\url{http://www.lirmm.fr/~abdaoui/FEEL.html}
}
\usage{
sentiments_score
}
\description{
Dataset from: Amine Abdaoui, Jérôme Azé, Sandra Bringay et Pascal Poncelet. FEEL: French Expanded Emotion Lexicon. Language Resources and Evaluation, LRE 2016, pp 1-23.
}
\keyword{datasets}
