% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_novels.R
\docType{data}
\name{sodomeetgomorrhe}
\alias{sodomeetgomorrhe}
\title{Marcel Proust's novel "Sodome et Gomorrhe"}
\format{A tibble with text, book, volume, and year}
\source{
<https://fr.wikisource.org/wiki/Sodome_et_Gomorrhe>
}
\usage{
sodomeetgomorrhe
}
\description{
A dataset containing Marcel Proust's "Sodom et Gomorrhe". 
This text has been downloaded from WikiSource.
}
\keyword{datasets}
