% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{sodomeetgomorrhe1}
\alias{sodomeetgomorrhe1}
\title{The first part of Marcel Proust's novel "Sodome et Gomorrhe"}
\format{A character vector with 448 elements}
\source{
\url{https://beq.ebooksgratuits.com/auteurs/Proust/proust.htm}
}
\usage{
sodomeetgomorrhe1
}
\description{
A dataset containing the first part of Marcel Proust's "À l’ombre des jeunes filles en fleurs". 
This text has been downloaded and converted to txt from "La Bibliothèque électronique du Québec".
}
\keyword{datasets}
