% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_datasets.R
\docType{data}
\name{stop_words}
\alias{stop_words}
\title{Stopwords}
\format{A tibble}
\source{
\url{https://raw.githubusercontent.com/stopwords-iso/stopwords-iso/master/stopwords-iso.json}
}
\usage{
stop_words
}
\description{
ISO stopwords
}
\keyword{datasets}
