% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_datasets.R
\docType{data}
\name{stop_words_fr}
\alias{stop_words_fr}
\title{Stop words}
\format{A tibble}
\source{
\url{https://raw.githubusercontent.com/stopwords-iso/stopwords-fr/master/stopwords-fr.txt}

\url{http://www.ranks.nl/stopwords/french}

\url{http://www.naunaute.com/liste-stop-words-francais-393}

\url{https://sites.google.com/site/kevinbouge/stopwords-lists/stopwords_fr.txt?attredirects=0&d=1}
}
\usage{
stop_words_fr
}
\description{
Stop words concatenated from various web sources.
}
\keyword{datasets}
