% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_norm.R
\name{unacent}
\alias{unacent}
\title{Remove accents}
\usage{
unacent(text)
}
\arguments{
\item{text}{a vector}
}
\value{
a vector
}
\description{
Remove accents from a character vector
}
\examples{
unacent("l'été")
}
