% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_novels.R
\docType{data}
\name{albertinedisparue}
\alias{albertinedisparue}
\title{Marcel Proust's novel "Albertine disparue"}
\format{A tibble with text, book, volume, and year}
\source{
<https://fr.wikisource.org/wiki/Albertine_disparue>
}
\usage{
albertinedisparue
}
\description{
A dataset containing Marcel Proust's "Albertine disparue". 
This text has been downloaded from WikiSource.
}
\keyword{datasets}
