% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_novels.R
\docType{data}
\name{laprisonniere}
\alias{laprisonniere}
\title{Marcel Proust's novel "La Prisonnière"}
\format{A tibble with text, book, volume, and year}
\source{
<https://fr.wikisource.org/wiki/La_Prisonni%C3%A8re>
}
\usage{
laprisonniere
}
\description{
A dataset containing Marcel Proust's "La prisonnière". 
This text has been downloaded from WikiSource.
}
\keyword{datasets}
