% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_books.R
\name{proust_books}
\alias{proust_books}
\title{Tidy data frame of Marcel Proust's 7 novels from La Recherche}
\usage{
proust_books()
}
\value{
A tibble with four columns: \code{text}, \code{book}, \code{volume} and \code{year}.
}
\description{
Returns a tidy tibble of Marcel Proust's 7 novels from À la recherche du temps
perdu. The tibble contains four columns: text, book, volume and year.
}
\examples{

#Create the tibble 
proust <- proust_books()
 

}
