% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_char.R
\name{proust_characters}
\alias{proust_characters}
\title{Characters from Proust Books}
\source{
\url{http://proust-personnages.fr/}
}
\usage{
proust_characters()
}
\value{
A tibble
}
\description{
Returns a tidy data frame of Marcel Proust's characters.
}
\examples{

#Creates the tibble 
proust <- proust_characters()
 

}
