% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_sentiments.R
\name{proust_sentiments}
\alias{proust_sentiments}
\title{Old sentiment lexicon
This function has been deprecated, and will be in next proustr version.
See the {rfeel} package now: http://github.com/ColinFay/rfeel}
\usage{
proust_sentiments(type = c("polarity", "score"))
}
\arguments{
\item{type}{For backward compatibility}
}
\value{
a tibble
}
\description{
Old sentiment lexicon
This function has been deprecated, and will be in next proustr version.
See the {rfeel} package now: http://github.com/ColinFay/rfeel
}
