% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_sentiments.R
\name{proust_sentiments}
\alias{proust_sentiments}
\title{Sentiment Lexicon}
\source{
Amine Abdaoui, Jérôme Azé, Sandra Bringay et Pascal Poncelet. FEEL: French Expanded Emotion Lexicon. Language Resources and Evaluation, LRE 2016, pp 1-23.
}
\usage{
proust_sentiments(type = "polarity")
}
\arguments{
\item{type}{polarity (positive or negative) or score on six sentiments (joy, fear, sadness, anger, surprise, disgust)}
}
\value{
a tibble
}
\description{
A sentiment lexicon with eitheir polarity or score.
}
\examples{
proust_sentiments(type = "score")
}
