% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proust_stop_words.R
\name{proust_stopwords}
\alias{proust_stopwords}
\title{Stop Words}
\source{
\url{https://raw.githubusercontent.com/stopwords-iso/stopwords-fr/master/stopwords-fr.txt}
}
\usage{
proust_stopwords()
}
\value{
a tibble with stopwords
}
\description{
Stop words concatenated from various web sources.
}
\examples{
proust_stopwords()
}
